zapomniałem dodać to stage
