+++ Party_Animals/src/main/resources/simulationwindow.fxml
@@ -9,12 +9,10 @@
 <?import java.net.URL?>
 <BorderPane xmlns="http://javafx.com/javafx"
             xmlns:fx="http://javafx.com/fxml"
-            fx:controller="agh.ics.oop.presenter.SimulationPresenter"
+            fx:controller="agh.ics.oop.presenter.SimulationWindowPresenter"
             prefHeight="400.0" prefWidth="600.0"
             styleClass="border-pane">
-    <stylesheets>
-        <URL value="@style.css" />
-    </stylesheets>
+
     <center>
         <VBox styleClass="centerBar">
             <Label fx:id="infoLabel" text="All animals will be living here!" textAlignment="CENTER"/>
